requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
discord.py==1.4.1
idna==2.10
keyboard==0.13.5
multidict==4.7.6
numpy==1.19.2
yarl==1.5.1
